fix: New architecture + imports fixed
